+++ Realm/RLMNetworkClient.h
@@ -22,36 +22,43 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+@interface RLMNetworkRequestOptions : NSObject
+@property (nonatomic, copy, nullable) NSString *authorizationHeaderName;
+@property (nonatomic, copy, nullable) NSDictionary<NSString *, NSString *> *customHeaders;
+@end
+
 /// An abstract class representing a server endpoint.
 @interface RLMSyncServerEndpoint : NSObject RLM_SYNC_UNINITIALIZABLE
++ (instancetype)endpoint;
+
++ (void)sendRequestToServer:(NSURL *)serverURL
+                       JSON:(NSDictionary *)jsonDictionary
+                    options:(nullable RLMNetworkRequestOptions *)options
+                 completion:(void (^)(NSError *))completionBlock;
 @end
 
 /// The authentication endpoint.
 @interface RLMSyncAuthEndpoint : RLMSyncServerEndpoint RLM_SYNC_UNINITIALIZABLE
-+ (instancetype)endpoint;
 @end
 
 /// The password change endpoint.
 @interface RLMSyncChangePasswordEndpoint : RLMSyncServerEndpoint RLM_SYNC_UNINITIALIZABLE
-+ (instancetype)endpoint;
+@end
+
+@interface RLMSyncUpdateAccountEndpoint : RLMSyncServerEndpoint RLM_SYNC_UNINITIALIZABLE
 @end
 
 /// The get user info endpoint.
 @interface RLMSyncGetUserInfoEndpoint : RLMSyncServerEndpoint RLM_SYNC_UNINITIALIZABLE
-+ (instancetype)endpoint;
-@end
-
-@interface RLMNetworkRequestOptions : NSObject
-
-@property (nonatomic, copy, nullable) NSString *authorizationHeaderName;
-@property (nonatomic, copy, nullable) NSDictionary<NSString *, NSString *> *customHeaders;
-
 @end
 
 /**
  A simple Realm Object Server network client that wraps `NSURLSession`.
  */
 @interface RLMNetworkClient : NSObject
+
+// Set the timeout in seconds for requests which do not take an explicit timeout.
++ (void)setDefaultTimeout:(NSTimeInterval)timeout;
 
 + (void)sendRequestToEndpoint:(RLMSyncServerEndpoint *)endpoint
                        server:(NSURL *)serverURL